Fixed bug with directory on 64bit
git-svn-id: https://vimpc.svn.sourceforge.net/svnroot/vimpc/trunk@654 fc90b29e-3040-4360-8be2-b397c2834534

+++ src/buffer/directory.cpp
@@ -148,7 +148,7 @@ void Directory::AddChild(std::string directory)
    std::string Parent = "";
    std::string Path = directory;
 
-   uint32_t i = 0;
+   size_t i = 0;
 
    while (i != std::string::npos)
    {
